+++ src/main/java/de/codescape/jira/plugins/scrumpoker/service/ScrumPokerSettingServiceImpl.java
@@ -3,6 +3,7 @@ package de.codescape.jira.plugins.scrumpoker.service;
 import com.atlassian.activeobjects.external.ActiveObjects;
 import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerSetting;
 import de.codescape.jira.plugins.scrumpoker.model.AllowRevealDeck;
+import de.codescape.jira.plugins.scrumpoker.model.DisplayCommentsForIssue;
 import de.codescape.jira.plugins.scrumpoker.model.GlobalSettings;
 import net.java.ao.DBParam;
 import net.java.ao.Query;
@@ -23,6 +24,7 @@ public class ScrumPokerSettingServiceImpl implements ScrumPokerSettingService {
     private static final String ALLOW_REVEAL_DECK = "allowRevealDeck";
     private static final String DISPLAY_DROPDOWN_ON_BOARDS = "displayDropdownOnBoards";
     private static final String CHECK_PERMISSION_TO_SAVE_ESTIMATE = "checkPermissionToSaveEstimate";
+    private static final String DISPLAY_COMMENTS_FOR_ISSUE = "displayCommentsForIssue";
 
     private final ActiveObjects activeObjects;
 
@@ -34,12 +36,20 @@ public class ScrumPokerSettingServiceImpl implements ScrumPokerSettingService {
     @Override
     public GlobalSettings load() {
         GlobalSettings globalSettings = new GlobalSettings();
-        globalSettings.setStoryPointField(loadString(STORY_POINT_FIELD, null));
-        globalSettings.setSessionTimeout(loadInteger(SESSION_TIMEOUT, SESSION_TIMEOUT_DEFAULT));
-        globalSettings.setAllowRevealDeck(AllowRevealDeck.valueOf(loadString(ALLOW_REVEAL_DECK, ALLOW_REVEAL_DECK_DEFAULT.name())));
-        globalSettings.setDefaultProjectActivation(loadBoolean(DEFAULT_PROJECT_ACTIVATION, DEFAULT_PROJECT_ACTIVATION_DEFAULT));
-        globalSettings.setDisplayDropdownOnBoards(loadBoolean(DISPLAY_DROPDOWN_ON_BOARDS, DISPLAY_DROPDOWN_ON_BOARDS_DEFAULT));
-        globalSettings.setCheckPermissionToSaveEstimate(loadBoolean(CHECK_PERMISSION_TO_SAVE_ESTIMATE, CHECK_PERMISSION_TO_SAVE_ESTIMATE_DEFAULT));
+        globalSettings.setStoryPointField(
+            loadString(STORY_POINT_FIELD, null));
+        globalSettings.setSessionTimeout(
+            loadInteger(SESSION_TIMEOUT, SESSION_TIMEOUT_DEFAULT));
+        globalSettings.setAllowRevealDeck(
+            AllowRevealDeck.valueOf(loadString(ALLOW_REVEAL_DECK, ALLOW_REVEAL_DECK_DEFAULT.name())));
+        globalSettings.setDefaultProjectActivation(
+            loadBoolean(DEFAULT_PROJECT_ACTIVATION, DEFAULT_PROJECT_ACTIVATION_DEFAULT));
+        globalSettings.setDisplayDropdownOnBoards(
+            loadBoolean(DISPLAY_DROPDOWN_ON_BOARDS, DISPLAY_DROPDOWN_ON_BOARDS_DEFAULT));
+        globalSettings.setCheckPermissionToSaveEstimate(
+            loadBoolean(CHECK_PERMISSION_TO_SAVE_ESTIMATE, CHECK_PERMISSION_TO_SAVE_ESTIMATE_DEFAULT));
+        globalSettings.setDisplayCommentsForIssue(DisplayCommentsForIssue.valueOf(
+            loadString(DISPLAY_COMMENTS_FOR_ISSUE, DISPLAY_COMMENTS_FOR_ISSUE_DEFAULT.name())));
         return globalSettings;
     }
 
@@ -51,6 +61,7 @@ public class ScrumPokerSettingServiceImpl implements ScrumPokerSettingService {
         persist(ALLOW_REVEAL_DECK, globalSettings.getAllowRevealDeck().name());
         persist(DISPLAY_DROPDOWN_ON_BOARDS, String.valueOf(globalSettings.isDisplayDropdownOnBoards()));
         persist(CHECK_PERMISSION_TO_SAVE_ESTIMATE, String.valueOf(globalSettings.isCheckPermissionToSaveEstimate()));
+        persist(DISPLAY_COMMENTS_FOR_ISSUE, globalSettings.getDisplayCommentsForIssue().name());
     }
 
     private String loadString(String key, String defaultValue) {